# src/main/resources/documentation/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Demo Ecommerce API",
    "description": "API documentation for managing products, sales and users",
    "contact": {
      "name": "Diego Pimenta",
      "url": "https://github.com/Diego-Pimenta",
      "email": "diegopimenta888@gmail.com"
    },
    "license": {
      "name": "Apache License",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "v1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/v1/auth/login": {
      "post": {
        "tags": [
          "Post",
          "Authentication"
        ],
        "summary": "Authenticate an user",
        "description": "Generates a jwt token with permissions to the user",
        "operationId": "authentication",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthenticationRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "User successfully authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid user credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardError"
                }
              }
            }
          },
          "401": {
            "description": "User credentials don't match",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardError"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "StandardError": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "path": {
            "type": "string"
          }
        }
      },
...
      "AuthenticationResponseDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "expiration": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "AuthenticationRequestDto": {
        "required": [
          "cpf",
          "password"
        ],
        "type": "object",
        "properties": {
          "cpf": {
            "maxLength": 11,
            "minLength": 11,
            "type": "string"
          },
          "password": {
            "maxLength": 20,
            "minLength": 8,
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "security": {
        "type": "http",
        "description": "Authentication required with bearer token",
        "name": "security",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}